savepoint update another move item call to check for target container's avalible capacity updated groupIDs from uint32 to uint16 (trying to save a bit more memory) updated missile code for new effects class and module rewrites implemented weapons in active module added proper checks in generic module corrected some module manager problems updated other weapon code to support new effects and updated modules
still in testing phase

+++ src/eve-server/inventory/InventoryItem.h
@@ -87,8 +87,8 @@ public:
     const std::string &     customInfo() const          { return m_customInfo; }
 
     /* public type queries  */
-    uint32                  typeID() const              { return m_type.id(); }
-    uint32                  groupID() const             { return m_type.groupID(); }
+    uint16                  typeID() const              { return m_type.id(); }
+    uint16                  groupID() const             { return m_type.groupID(); }
     double                  radius() const              { return (HasAttribute(AttrRadius) ? GetAttribute(AttrRadius).get_double() : 1.0); }
     const ItemGroup &       group() const               { return m_type.group(); }
     const ItemCategory &    category() const            { return m_type.category(); }
